requirements.txt
Appium-Python-Client==3.1.0
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
h11==0.14.0
idna==3.6
numpy==1.26.2
opencv-python==4.8.1.78
outcome==1.3.0.post0
pycparser==2.21
PySocks==1.7.1
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
